Change Data Type
Because firestore stores numbers as Long

+++ RoadReport/app/src/main/java/com/txwstudio/app/roadreport/activity/AccidentEventActivity.kt
@@ -84,7 +84,7 @@ class AccidentEventActivity : AppCompatActivity() {
             FirebaseAuthHelper().getCurrUserName(),
             FirebaseAuthHelper().getCurrUserUid(),
             Timestamp(Date()),
-            situationType,
+            situationType.toLong(),
             editText_accidentEvent_locationContent.text.toString(),
             editText_accidentEvent_situationContent.text.toString()
         )
